docs: add README with links
