+++ README.md
@@ -1,38 +1,13 @@
-# React-shop-cloudfront
+#### This is main directory
 
-This is frontend starter project for nodejs-aws mentoring program. It uses the following technologies:
+##### At /webApp you find all files related to React + Vite web app
 
-- [Vite](https://vitejs.dev/) as a project bundler
-- [React](https://beta.reactjs.org/) as a frontend framework
-- [React-router-dom](https://reactrouterdotcom.fly.dev/) as a routing library
-- [MUI](https://mui.com/) as a UI framework
-- [React-query](https://react-query-v3.tanstack.com/) as a data fetching library
-- [Formik](https://formik.org/) as a form library
-- [Yup](https://github.com/jquense/yup) as a validation schema
-- [Vitest](https://vitest.dev/) as a test runner
-- [MSW](https://mswjs.io/) as an API mocking library
-- [Eslint](https://eslint.org/) as a code linting tool
-- [Prettier](https://prettier.io/) as a code formatting tool
-- [TypeScript](https://www.typescriptlang.org/) as a type checking tool
+##### At /webApp/dist - build files fro static web site
 
-## Available Scripts
+##### At /cdk - all files for configuring AWS CDK Constructs and Stacks
 
-### `start`
+#### Task - 2
 
-Starts the project in dev mode with mocked API on local environment.
+CloudFront URL: https://d1ta9wgqox65lg.cloudfront.net/ - works
 
-### `build`
-
-Builds the project for production in `dist` folder.
-
-### `preview`
-
-Starts the project in production mode on local environment.
-
-### `test`, `test:ui`, `test:coverage`
-
-Runs tests in console, in browser or with coverage.
-
-### `lint`, `prettier`
-
-Runs linting and formatting for all files in `src` folder.
+S3-website: https://ts-react-shop-cdk-s3.s3.amazonaws.com/index.html - <Message>Access Denied</Message>
