+++ webApp/src/components/PaperLayout/PaperLayout.tsx
