+++ webApp/src/mocks/browser.ts
