+++ webApp/src/models/Product.ts
