+++ webApp/src/utils/utils.ts
